// post-update/zenstack/schema.ts
export const schema = {
    provider: {
        type: "sqlite"
    },
    models: {
        User: {
            name: "User",
            fields: {
                id: {
                    name: "id",
                    type: "Int",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("autoincrement") }] }],
                    default: ExpressionUtils.call("autoincrement")
                },
                email: {
                    name: "email",
                    type: "String",
                    unique: true,
                    attributes: [{ name: "@unique" }]
                },
                posts: {
                    name: "posts",
                    type: "Post",
                    array: true,
                    relation: { opposite: "author" }
                }
            },
            attributes: [
                { name: "@@allow", args: [{ name: "operation", value: ExpressionUtils.literal("all") }, { name: "condition", value: ExpressionUtils.literal(true) }] }
            ],
            idFields: ["id"],
            uniqueFields: {
                id: { type: "Int" },
                email: { type: "String" }
            }
        },
        Post: {
            name: "Post",
            fields: {
                id: {
                    name: "id",
                    type: "Int",
                    id: true,
                    attributes: [{ name: "@id" }, { name: "@default", args: [{ name: "value", value: ExpressionUtils.call("autoincrement") }] }],
                    default: ExpressionUtils.call("autoincrement")
                },
                title: {
                    name: "title",
                    type: "String"
                },
                published: {
                    name: "published",
                    type: "Boolean",
                    attributes: [{ name: "@default", args: [{ name: "value", value: ExpressionUtils.literal(false) }] }],
                    default: false
                },
                author: {
                    name: "author",
                    type: "User",
                    optional: true,
                    attributes: [{ name: "@relation", args: [{ name: "fields", value: ExpressionUtils.array([ExpressionUtils.field("authorId")]) }, { name: "references", value: ExpressionUtils.array([ExpressionUtils.field("id")]) }] }],
                    relation: { opposite: "posts", fields: ["authorId"], references: ["id"] }
                },
                authorId: {
                    name: "authorId",
                    type: "Int",
                    optional: true,
                    foreignKeyFor: [
                        "author"
                    ]
                }
            },
            attributes: [
                { name: "@@allow", args: [{ name: "operation", value: ExpressionUtils.literal("all") }, { name: "condition", value: ExpressionUtils.literal(true) }] },
                { name: "@@deny", args: [{ name: "operation", value: ExpressionUtils.literal("post-update") }, { name: "condition", value: ExpressionUtils.binary(ExpressionUtils.field("published"), "&&", ExpressionUtils.binary(ExpressionUtils.member(ExpressionUtils.call("auth"), ["id"]), "!=", ExpressionUtils.field("authorId"))) }] },
                { name: "@@deny", args: [{ name: "operation", value: ExpressionUtils.literal("post-update") }, { name: "condition", value: ExpressionUtils.binary(ExpressionUtils.field("authorId"), "!=", ExpressionUtils.member(ExpressionUtils.call("before"), ["authorId"])) }] }
            ],
            idFields: ["id"],
            uniqueFields: {
                id: { type: "Int" }
            }
        }
    },
    authType: "User",
    plugins: {}
} as const satisfies SchemaDef;